core, eth/protocols/snap: Snap/2 Protocol + BAL Serving (#34083)
Implement the snap/2 wire protocol with BAL serving

---------
